created Review constructor, changed renderReviews function
