+++ src/game/index.js
